Feature/Feature Pokemon pagination (#19)
* [Feat] Added permission of Internet

* [Feat] Added retrofit, gson, paging 3 and coil dependencies

* [Feat] Added provides of Hilt

* [Feat] Changed the package path

* [Feat] Added rules in proguard for DTOs and retrofit

* [Feat] Added new versions of dependencies

* [Feat] Added loading animation

* [Feat] Added PokemonScreen in Navigation

* [Feat] Added PokemonApi

* [Feat] Added PokemonDataSource

* [Feat] Added models of Pokemon

* [Feat] Added repository and mappers for Pokemons models

* [Feat] Added utils functions in Resource

* [Feat] Added new strings.xml

* [Feat] Added PokemonPagingSource and PokemonScreen

* [Feat] Added PokemonViewModel

* [Feat] Added DomainContract

* [Feat] Added shimmer library

* [Feat] Added shimmer in PokemonScreen
