+++ app/src/main/java/com/dgalan/pokeapp/authetication/data/repository/FirebaseAuthRepositoryContract.kt
@@ -1,0 +1,10 @@
+package com.dgalan.pokeapp.authetication.data.repository
+
+import com.dgalan.pokeapp.utils.state.Resource
+import com.google.firebase.auth.FirebaseUser
+
+interface FirebaseAuthRepositoryContract {
+
+    suspend fun loginUserInFirebase(email: String, password: String): Resource<FirebaseUser>
+    suspend fun registerUserInFirebase(name: String, email: String, password: String): Resource<FirebaseUser>
+}
