+++ app/src/main/java/com/dgalan/pokeapp/authetication/ui/login/state/LoginUIEvent.kt
@@ -1,0 +1,9 @@
+package com.dgalan.pokeapp.authetication.ui.login.state
+
+sealed class LoginUIEvent {
+    data class EmailChanged(val email: String) : LoginUIEvent()
+    data class PasswordChanged(val password: String) : LoginUIEvent()
+    data class PasswordVisibilityChanged(val isPasswordVisible: Boolean) : LoginUIEvent()
+    data object LoginButtonClicked : LoginUIEvent()
+    data object ResetResourceState : LoginUIEvent()
+}
