+++ app/src/main/java/com/dgalan/pokeapp/authetication/ui/login/state/LoginUIState.kt
@@ -1,0 +1,7 @@
+package com.dgalan.pokeapp.authetication.ui.login.state
+
+data class LoginUIState(
+    val email: String = "",
+    val password: String = "",
+    val isPasswordVisible: Boolean = false
+)
