--- app/src/main/java/com/dgalan/pokeapp/login/ui/state/LoginUIEvent.kt
@@ -1,9 +1,0 @@
-package com.dgalan.pokeapp.login.ui.state
-
-sealed class LoginUIEvent {
-    data class EmailChanged(val email: String) : LoginUIEvent()
-    data class PasswordChanged(val password: String) : LoginUIEvent()
-    data class PasswordVisibilityChanged(val isPasswordVisible: Boolean) : LoginUIEvent()
-    data object LoginButtonClicked : LoginUIEvent()
-    data object ResetResourceState : LoginUIEvent()
-}
