--- app/src/main/java/com/dgalan/pokeapp/login/ui/state/LoginUIState.kt
@@ -1,7 +1,0 @@
-package com.dgalan.pokeapp.login.ui.state
-
-data class LoginUIState(
-    val email: String = "",
-    val password: String = "",
-    val isPasswordVisible: Boolean = false
-)
