+++ app/src/main/java/com/dgalan/pokeapp/pokemons/data/datasource/PokemonDataSourceContract.kt
@@ -1,0 +1,13 @@
+package com.dgalan.pokeapp.pokemons.data.datasource
+
+import com.dgalan.pokeapp.pokemons.data.model.PokemonDTO
+
+interface PokemonDataSourceContract {
+    interface Remote {
+
+        suspend fun getPokemonList(page: Int): PokemonDTO
+    }
+
+    interface Local {
+    }
+}
