+++ app/src/main/java/com/dgalan/pokeapp/pokemons/domain/DomainContract.kt
@@ -1,0 +1,14 @@
+package com.dgalan.pokeapp.pokemons.domain
+
+import com.dgalan.pokeapp.pokemons.domain.model.Pokemon
+
+interface DomainContract {
+    interface Repository {
+
+        suspend fun getPokemonList(page: Int): Pokemon
+    }
+
+    interface UseCase {
+
+    }
+}
