+++ app/src/main/java/com/dgalan/pokeapp/pokemons/domain/model/Pokemon.kt
@@ -1,0 +1,13 @@
+package com.dgalan.pokeapp.pokemons.domain.model
+
+data class Pokemon(
+    val count: Int,
+    val next: String,
+    val previous: String?,
+    val results: List<PokemonResult>
+)
+
+data class PokemonResult(
+    val name: String,
+    val url: String
+)
