+++ app/src/main/java/com/dgalan/pokeapp/ui/navigation/Screens.kt
@@ -3,4 +3,5 @@ package com.dgalan.pokeapp.ui.navigation
 sealed class Screens(val route: String) {
     data object LoginScreen : Screens("login_screen")
     data object RegisterScreen : Screens("register_screen")
+    data object PokemonScreen : Screens("pokemon_screen")
 }
