+++ app/src/main/java/com/dgalan/pokeapp/utils/di/CoroutineDispatcherModule.kt
@@ -7,24 +7,29 @@ import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Qualifier
+import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 object CoroutineDispatcherModule {
 
     @IoDispatcher
+    @Singleton
     @Provides
     fun providesIoDispatcher(): CoroutineDispatcher = Dispatchers.IO
 
     @DefaultDispatcher
+    @Singleton
     @Provides
     fun providesDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
 
     @MainDispatcher
+    @Singleton
     @Provides
     fun providesMainDispatcher(): CoroutineDispatcher = Dispatchers.Main
 
     @UnconfinedDispatcher
+    @Singleton
     @Provides
     fun providesUnconfinedDispatcher(): CoroutineDispatcher = Dispatchers.Unconfined
