--- app/src/main/java/com/dgalan/pokeapp/utils/di/FirebaseAuthModule.kt
@@ -1,22 +1,0 @@
-package com.dgalan.pokeapp.utils.di
-
-import com.dgalan.pokeapp.firebasedata.FirebaseAuthRepositoryContract
-import com.dgalan.pokeapp.firebasedata.repository.FirebaseAuthRepository
-import com.google.firebase.auth.FirebaseAuth
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@InstallIn(SingletonComponent::class)
-@Module
-class FirebaseAuthModule {
-
-    @Provides
-    fun provideFirebaseAuth(): FirebaseAuth =
-        FirebaseAuth.getInstance()
-
-    @Provides
-    fun provideFirebaseAuthRepository(firebaseAuthRepository: FirebaseAuthRepository): FirebaseAuthRepositoryContract =
-        firebaseAuthRepository
-}
