+++ app/src/main/res/drawable/ic_pokeball_byn.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="275dp"
+    android:height="276dp"
+    android:viewportHeight="276"
+    android:viewportWidth="275">
+  <path
+      android:fillColor="#424242"
+      android:pathData="M122.66,0.5C91.15,4.07 62.3,17.94 39.94,40.31C24.31,55.87 13.29,73.7 6.29,94.64C4.08,101.19 1.23,113.64 0.45,120.25L0,123.62L34.81,123.49L69.63,123.29L70.86,118.76C73.65,108.51 80.52,97.04 88.56,88.93C112.87,64.69 150.99,61.83 178.74,82.32C190.67,91.07 200.52,105.47 204.09,119.21L205.19,123.62H240.07H274.95L274.56,120.83C269.89,85.95 252.91,54.51 226.98,32.66C207.72,16.45 185.29,6.01 160.52,1.67C151.51,0.12 131.48,-0.53 122.66,0.5Z" />
+  <path
+      android:fillColor="#424242"
+      android:pathData="M130.57,97.04C115.53,99.76 102.95,110.72 98.16,125.18C96.73,129.65 96.4,131.53 96.4,137.88C96.4,144.24 96.73,146.12 98.16,150.59C102.37,163.23 111.83,172.7 124.87,177.17C129.34,178.73 131.09,178.99 137.51,178.99C143.93,178.99 145.68,178.73 150.15,177.17C162.34,172.96 171.74,164.07 176.08,152.53C184.32,130.56 172.65,106.18 150.35,98.72C144.7,96.78 136.02,96.07 130.57,97.04Z" />
+  <path
+      android:fillColor="#424242"
+      android:pathData="M0.45,155.52C0.65,157.4 1.56,162.45 2.46,166.73C10.05,203.49 33.65,236.56 66.19,256.14C115.72,285.96 178.55,281.74 223.8,245.7C251.48,223.59 269.7,191.04 274.56,154.87L274.95,152.15H240.07H205.19L204.09,156.49C198.77,177.04 181.98,195.26 161.43,202.78C152.74,205.96 147.82,206.8 137.83,206.86C128.24,206.93 124.28,206.28 115.79,203.56C94.65,196.75 76.63,178.14 70.86,157.01L69.63,152.47L34.81,152.27L0,152.15L0.45,155.52Z" />
+</vector>
