+++ app/src/main/res/drawable/ic_pokemon_silhouette.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="1280dp"
+    android:height="1173dp"
+    android:viewportHeight="1173"
+    android:viewportWidth="1280">
+  <path
+      android:fillColor="#000000"
+      android:pathData="M0.3,1.6c-0.9,2.4 23.8,51.3 36.7,72.4 16.4,27 19.6,31.2 50.7,67.7 29.8,34.8 43.4,48.7 71.3,72.4 27.5,23.3 51.3,40.2 85.5,60.8 9.4,5.7 17.1,10.6 17.3,11.1 0.2,0.5 -1.2,3.9 -3.1,7.7 -12,24.1 -19.5,56.3 -25.6,110.3 -3.2,28.1 -7.6,44.2 -19.1,69.5 -3.8,8.2 -7.6,17.7 -8.6,21 -2.1,7.5 -1.4,18.4 1.7,27.5 4.1,11.8 20.4,43 27.3,52 8.4,11.1 27.8,32.8 36.4,40.7 5.7,5.4 6.5,6.5 5.6,8 -21.2,33.2 -29.2,50 -35.4,74.8 -4.8,18.7 -6.3,34.1 -6.4,62 0,23.9 0.1,25.4 2.6,34.5 6.4,23.5 19.6,52.1 34.8,75.8 7.9,12.2 9.1,14.7 8.6,17.3 -1,5.1 -5.6,15.8 -14.3,33.2 -15.7,31.4 -20.3,45.9 -21.1,65.4 -0.6,15.5 1.2,23.9 7.9,37.2 8.8,17.6 23,35.3 39.7,49.2l4.2,3.6 -5.1,6.4c-16.4,20.5 -27.9,42.7 -27.9,54 0,4.4 1.7,7.9 3.9,7.9 0.3,-0 1.2,1.4 1.8,3 0.9,2.4 1.7,3 4,3 1.5,-0 4.1,0.7 5.7,1.5 1.5,0.8 3.5,1.5 4.2,1.5 2.4,-0 23.5,-11.3 33.4,-18 11.9,-7.9 30.8,-21.8 38,-28 11.6,-9.8 10.1,-9.3 31.4,-10.1 10.5,-0.5 31,-1.3 45.6,-1.9 18.1,-0.7 31.7,-1.8 43,-3.5 9.1,-1.4 18.6,-2.5 21.2,-2.5 8.6,-0 29.2,3.2 55.8,8.6 58.1,12 65.6,13.5 75.1,14.7l9.8,1.2 -5.3,12.2c-3,6.7 -6.4,15 -7.6,18.5 -2.7,7.9 -3,16.1 -0.5,18.8 2.7,3 13.4,8 17.1,8 1.8,-0 3.6,0.5 3.9,1 1.8,3 13,-1.7 29.6,-12.5 19.5,-12.7 29.2,-20.7 48.5,-40.1 15.6,-15.7 20.3,-19.8 25.5,-22.4 12.8,-6.4 23.3,-14.1 33.4,-24.4 11.8,-11.9 17.9,-20.9 30.5,-44.4 21.4,-40.3 31.1,-70.1 31.6,-97.5l0.2,-12.8 18.4,-6.4c10,-3.5 26.2,-9.2 35.8,-12.8 9.6,-3.5 20,-7.2 23,-8.2 6,-1.9 9,-4 9,-6.2 0,-0.8 -3,-8.1 -6.7,-16.1 -8.5,-18.9 -14.9,-34.2 -16.2,-38.7 -1.2,-4.4 -4,-2.8 36.2,-19.7 17.2,-7.1 34.1,-14.4 37.7,-16.3 9.2,-4.6 9.8,-0.4 -10.1,-65.3 -5.4,-17.7 -9.9,-32.9 -9.9,-33.6 0,-0.8 1,-1.6 2.3,-1.9 1.2,-0.3 10.5,-3.5 20.7,-7.2 10.2,-3.6 27.5,-9.8 38.5,-13.7 11,-3.9 22.8,-8.3 26.1,-9.8l6.2,-2.8 -0.9,-4.8c-0.5,-2.7 -2.7,-10.1 -4.8,-16.4 -2.2,-6.3 -9.2,-28.6 -15.5,-49.5 -6.4,-20.9 -12.7,-41.5 -14,-45.7l-2.4,-7.7 3.1,-1c1.8,-0.5 12,-3.3 22.7,-6.1 10.7,-2.9 28.1,-7.8 38.5,-11.1 22.3,-6.8 160.5,-53 175,-58.4 13.2,-5 23.5,-10.2 23.5,-12 0,-5.2 -10.1,-30.9 -23.3,-59 -6.5,-13.9 -54.4,-111.7 -75.8,-154.8l-7.3,-14.9 -3.3,0.5c-4.4,0.7 -12.2,4.8 -52.4,27.4 -28.4,15.9 -51.8,30.6 -151.3,94.8 -64.7,41.8 -119.8,77.6 -122.4,79.5 -2.6,2 -4.8,4.5 -5,5.6 -0.2,1.2 2.8,6.3 7.4,12.9 12.4,17.8 118.4,177.5 120.5,181.8 1.1,2.1 1.9,4 1.7,4.2 -0.2,0.1 -4.6,1.7 -9.8,3.5 -5.2,1.8 -19.8,7.3 -32.5,12.2 -12.6,4.9 -28.8,11.2 -36,13.9 -7.1,2.8 -13.6,5.6 -14.4,6.4 -1.8,1.8 -0.8,5.4 6.1,22 14.5,34.5 32.8,80.2 32.8,81.7 0,1.1 -8.4,5.9 -28.8,16.3 -15.8,8.1 -29.3,15.2 -30,15.8 -2,1.6 -1,4.2 11.8,30.7 11.9,24.6 16,34 16,36.4 0,1.3 -35.2,18.9 -40.2,20.1 -3.3,0.8 -2.7,2.6 -7.8,-27 -4.2,-24.7 -8.9,-46.4 -16.5,-76 -4.2,-16.5 -16.2,-67.5 -23.2,-98 -11.7,-52 -28.1,-153.3 -34.8,-216 -9.8,-91.2 -10.5,-94.8 -29.1,-142.7 -4.6,-12 -8.4,-22.2 -8.4,-22.7 0,-0.5 2.1,-1.2 4.8,-1.5 11.7,-1.7 52.4,-13.3 83.7,-23.9 18,-6.1 59.1,-24.2 76,-33.4 28.2,-15.3 66.2,-40.5 91,-60.1 12.9,-10.3 56.4,-53.1 65.6,-64.7 10.9,-13.6 10.7,-14.5 -3.3,-14.5 -5.7,-0 -15.9,0.5 -22.8,1 -6.9,0.6 -22.8,1.4 -35.5,2 -34.3,1.5 -48.3,2.8 -65.5,6.1 -34.3,6.7 -85.6,20.6 -111.5,30.2 -47.2,17.7 -111,51.6 -146.8,78.1 -4.2,3.1 -8.6,5.6 -9.7,5.6 -1.1,-0 -9.4,-3.1 -18.4,-6.9 -36.3,-15.4 -70.3,-24.3 -107.6,-28.3 -18.1,-1.9 -60.9,-1.6 -79,0.5 -37.6,4.5 -65.1,11.3 -102.9,25.6 -7.4,2.8 -13.8,5.1 -14.1,5.1 -0.4,-0 -1.8,-1.5 -3.1,-3.4 -4.7,-6.6 -34.6,-39.3 -52.4,-57.2 -18.1,-18.1 -29.6,-28.2 -56.5,-49.5 -21.7,-17.3 -26.7,-20.8 -61,-42.6 -30.6,-19.5 -46.6,-28.9 -58,-34 -48.4,-21.7 -70.8,-30.7 -84,-33.7 -8.4,-2 -11.9,-2 -12.7,-0z"
+      android:strokeColor="#00000000" />
+</vector>
