+++ gradle/libs.versions.toml
@@ -3,12 +3,16 @@ activity-compose = "1.8.0"
 agp = "8.1.2"
 androidx-test-ext-junit = "1.1.5"
 ben-manes = "0.49.0"
+coilCompose = "2.5.0"
 comGoogleDevtoolsKspGradlePlugin = "1.9.10-1.0.13"
 compose-bom = "2023.10.01"
+composeShimmer = "1.2.0"
+converterGson = "2.9.0"
 core-ktx = "1.12.0"
 espresso-core = "3.5.1"
 firebaseBom = "32.4.0"
 googleServices = "4.4.0"
+gson = "2.10.1"
 hilt = "2.48.1"
 hiltNavigationCompose = "1.0.0"
 junit = "4.13.2"
@@ -19,6 +23,9 @@ material-icons-extended = "1.5.4"
 mockk = "1.13.8"
 navigationCompose = "2.7.4"
 org-jetbrains-kotlin-android = "1.9.10"
+pagingCompose = "3.2.1"
+pagingRuntimeKtx = "3.2.1"
+retrofit = "2.9.0"
 version-catalog-update = "0.8.1"
 
 [libraries]
@@ -27,13 +34,19 @@ androidx-hilt-navigation-compose = { module = "androidx.hilt:hilt-navigation-com
 androidx-lifecycle-runtime-compose = { module = "androidx.lifecycle:lifecycle-runtime-compose", version.ref = "lifecycle-runtime-ktx" }
 androidx-material-icons-extended = { module = "androidx.compose.material:material-icons-extended", version.ref = "material-icons-extended" }
 androidx-navigation-compose = { module = "androidx.navigation:navigation-compose", version.ref = "navigationCompose" }
+androidx-paging-compose = { module = "androidx.paging:paging-compose", version.ref = "pagingCompose" }
+androidx-paging-runtime-ktx = { module = "androidx.paging:paging-runtime-ktx", version.ref = "pagingRuntimeKtx" }
 androidx-test-ext-junit = { module = "androidx.test.ext:junit", version.ref = "androidx-test-ext-junit" }
+coil-compose = { module = "io.coil-kt:coil-compose", version.ref = "coilCompose" }
 com-google-devtools-ksp-gradle-plugin = { module = "com.google.devtools.ksp:com.google.devtools.ksp.gradle.plugin", version.ref = "comGoogleDevtoolsKspGradlePlugin" }
 com-google-firebase-auth-ktx = { module = "com.google.firebase:firebase-auth-ktx" }
 com-google-firebase-bom = { module = "com.google.firebase:firebase-bom", version.ref = "firebaseBom" }
 compose-bom = { module = "androidx.compose:compose-bom", version.ref = "compose-bom" }
+compose-shimmer = { module = "com.valentinilk.shimmer:compose-shimmer", version.ref = "composeShimmer" }
+converter-gson = { module = "com.squareup.retrofit2:converter-gson", version.ref = "converterGson" }
 core-ktx = { module = "androidx.core:core-ktx", version.ref = "core-ktx" }
 espresso-core = { module = "androidx.test.espresso:espresso-core", version.ref = "espresso-core" }
+gson = { module = "com.google.code.gson:gson", version.ref = "gson" }
 hilt-android = { module = "com.google.dagger:hilt-android", version.ref = "hilt" }
 hilt-android-compiler = { module = "com.google.dagger:hilt-android-compiler", version.ref = "hilt" }
 junit = { module = "junit:junit", version.ref = "junit" }
@@ -41,6 +54,7 @@ lifecycle-runtime-ktx = { module = "androidx.lifecycle:lifecycle-runtime-ktx", v
 lottie-compose = { module = "com.airbnb.android:lottie-compose", version.ref = "lottieCompose" }
 material3 = { module = "androidx.compose.material3:material3" }
 mockk = { module = "io.mockk:mockk", version.ref = "mockk" }
+retrofit = { module = "com.squareup.retrofit2:retrofit", version.ref = "retrofit" }
 ui = { module = "androidx.compose.ui:ui" }
 ui-graphics = { module = "androidx.compose.ui:ui-graphics" }
 ui-test-junit4 = { module = "androidx.compose.ui:ui-test-junit4" }
